refactor(tui): improve directory structure
