+++ src/interactive/clear_retained.rs
