+++ src/interactive/details/history/graph_data.rs
