+++ src/interactive/ui.rs
@@ -1,0 +1,9 @@
+use tui::style::Color;
+
+pub const fn focus_color(has_focus: bool) -> Color {
+    if has_focus {
+        Color::LightGreen
+    } else {
+        Color::Gray
+    }
+}
